move ModeSpec to file with all other XyzSpec types
